detect other encrypted dlinks

+++ modules/P11_dlink_SHRS_enc_extract.sh
@@ -21,13 +21,17 @@ P11_dlink_SHRS_enc_extract() {
   module_log_init "${FUNCNAME[0]}"
   NEG_LOG=0
 
-  if [[ "$DLINK_ENC_DETECTED" -eq 1 ]]; then
+  if [[ "$DLINK_ENC_DETECTED" -ne 0 ]]; then
     module_title "DLink encrypted firmware extractor"
     hexdump -C "$FIRMWARE_PATH" | head | tee -a "$LOG_FILE"
-    dd if="$FIRMWARE_PATH" skip=1756 iflag=skip_bytes|openssl aes-128-cbc -d -p -nopad -nosalt -K "c05fbf1936c99429ce2a0781f08d6ad8" -iv "67c6697351ff4aec29cdbaabf2fbe346" --nosalt -in /dev/stdin -out "$LOG_DIR"/firmware/firmware_dlink_dec.bin 2>&1 | tee -a "$LOG_FILE"
+    if [[ "$DLINK_ENC_DETECTED" -eq 1 ]]; then
+      dd if="$FIRMWARE_PATH" skip=1756 iflag=skip_bytes|openssl aes-128-cbc -d -p -nopad -nosalt -K "c05fbf1936c99429ce2a0781f08d6ad8" -iv "67c6697351ff4aec29cdbaabf2fbe346" --nosalt -in /dev/stdin -out "$LOG_DIR"/firmware/firmware_dlink_dec.bin 2>&1 | tee -a "$LOG_FILE"
+    elif [[ "$DLINK_ENC_DETECTED" -eq 2 ]]; then
+      print_output "[-] Decryption of this file is currently not supported"
+    fi
 
     if [[ -f "$LOG_DIR"/firmware/firmware_dlink_dec.bin ]]; then
-      print_output "[+] Decrypted D-Link firmware file to $LOG_DIR/firmware/firmware_dlink_dec.bin"
+      print_output "[+] Decrypted D-Link firmware file to $ORANGE$LOG_DIR/firmware/firmware_dlink_dec.bin$NC"
       export FIRMWARE_PATH="$LOG_DIR"/firmware/firmware_dlink_dec.bin
     else
       print_output "[-] Decryption of D-Link firmware file failed"
